requirements.txt
beautifulsoup4==4.4.1
mechanize==0.2.5
easygui==0.97.4
pycrypto==2.6.1
